lib: extract svg methods
