+++ src/simple-whiteboard.ts
@@ -7,55 +7,7 @@ import { customElement, state } from "lit/decorators.js";
 import { RoughCanvas } from "roughjs/bin/canvas";
 import { Options as RoughCanvasOptions } from "roughjs/bin/core";
 import { unsafeHTML } from "lit/directives/unsafe-html.js";
-
-/**
- * Calculate the average of two numbers.
- *
- * @param a The first number.
- * @param b The second number.
- * @returns The average of the two numbers.
- */
-const average = (a: number, b: number) => (a + b) / 2;
-
-/**
- * Get a SVG path from a stroke.
- *
- * @param points Coordinates of the points.
- * @param closed If the path should be closed.
- * @returns The SVG path.
- */
-const getSvgPathFromStroke = (points: number[][], closed = true) => {
-  const len = points.length;
-
-  if (len < 4) {
-    return ``;
-  }
-
-  let a = points[0];
-  let b = points[1];
-  const c = points[2];
-
-  let result = `M${a[0].toFixed(2)},${a[1].toFixed(2)} Q${b[0].toFixed(
-    2
-  )},${b[1].toFixed(2)} ${average(b[0], c[0]).toFixed(2)},${average(
-    b[1],
-    c[1]
-  ).toFixed(2)} T`;
-
-  for (let i = 2, max = len - 1; i < max; i++) {
-    a = points[i];
-    b = points[i + 1];
-    result += `${average(a[0], b[0]).toFixed(2)},${average(a[1], b[1]).toFixed(
-      2
-    )} `;
-  }
-
-  if (closed) {
-    result += "Z";
-  }
-
-  return result;
-};
+import { getSvgPathFromStroke } from "./lib/svg";
 
 type BoundingRect = {
   x: number;
